+++ CHANGELOG.md
@@ -6,6 +6,9 @@ _WIP_
 **Core**
 - Bump Go to 1.19
 
+**New features**
+- Collapse / uncollapse all replies with <kbd>h</kbd> / <kbd>l</kbd> 
+
 **Bugfixes**
 - Refreshing submissions no longer jumps to the first page just before fetching
 - Fixed a bug where submissions marked as read would not properly italicize after a refresh